requirements.txt
--extra-index-url https://erc.timetoact-group.at/
erc3>=1.2.0
langchain>=0.3.0
langchain-openai>=0.2.0
langchain-gigachat_lc1
python-dotenv>=1.0.0